erillinen dialogikomponentti, aikavalitstin
