+++ src/components/projekti/suunnitteluvaihe/SuunniteluvaiheenVuorovaikuttaminen.tsx
@@ -4,8 +4,7 @@ import SectionContent from "@components/layout/SectionContent";
 import { TallennaProjektiInput, Projekti, api, VuorovaikutusInput } from "@services/api";
 import Section from "@components/layout/Section";
 import { ReactElement, useEffect, useState } from "react";
-import { Badge, Chip, chipClasses, Stack } from "@mui/material";
-import { styled } from "@mui/material/styles";
+import { Stack } from "@mui/material";
 import Button from "@components/button/Button";
 import useSnackbars from "src/hooks/useSnackbars";
 import log from "loglevel";
@@ -16,14 +15,7 @@ import DatePicker from "@components/form/DatePicker";
 import dayjs from "dayjs";
 import { vuorovaikutusSchema } from "src/schemas/vuorovaikutus";
 import HassuStack from "@components/layout/HassuStack";
-import HassuDialog from "@components/HassuDialog";
-import WindowCloseButton from "@components/button/WindowCloseButton";
-import HeadphonesIcon from "@mui/icons-material/Headphones";
-import LocationCityIcon from "@mui/icons-material/LocationCity";
-import LocalPhoneIcon from "@mui/icons-material/LocalPhone";
-import TextInput from "@components/form/TextInput";
-import Select from "@components/form/Select";
-import HassuGrid from "@components/HassuGrid";
+import VuorovaikutusDialog from "./VuorovaikutustilaisuusDialog";
 
 type ProjektiFields = Pick<TallennaProjektiInput, "oid">;
 type RequiredProjektiFields = Required<{
@@ -64,16 +56,6 @@ export default function SuunniteluvaiheenVuorovaikuttaminen({
   const [isFormSubmitting, setIsFormSubmitting] = useState(false);
   const { showSuccessMessage, showErrorMessage } = useSnackbars();
   const today = dayjs().format();
-
-  const HassuBadge = styled(Badge)(() => ({
-    [`&.${chipClasses.deleteIcon}`]: {
-      marginLeft: "0.5rem",
-      marginRight: "1rem",
-    },
-    [`&.${chipClasses.label}`]: {
-      paddingRight: "0px",
-    },
-  }));
 
   const formOptions: UseFormProps<FormValues> = {
     resolver: yupResolver(vuorovaikutusSchema, { abortEarly: false, recursive: true }),
@@ -246,102 +228,7 @@ export default function SuunniteluvaiheenVuorovaikuttaminen({
           </Button>
         </Stack>
       </Section>
-      <div>
-        <HassuDialog
-          open={openVuorovaikutustilaisuus}
-          onClose={() => setOpenVuorovaikutustilaisuus(false)}
-          maxWidth={"lg"}
-        >
-          <Section noDivider smallGaps>
-            <SectionContent>
-              <div className="vayla-dialog-title flex">
-                <div className="flex-grow">Vuorovaikutustilaisuuden lisääminen</div>
-                <div className="justify-end">
-                  <WindowCloseButton onClick={() => setOpenVuorovaikutustilaisuus(false)}></WindowCloseButton>
-                </div>
-              </div>
-            </SectionContent>
-            <SectionContent>
-              <div className="vayla-dialog-content">
-                <form>
-                  <HassuStack>
-                    <p>Voit valita saman vuorovaikutustavan useammin kuin yhden kerran.</p>
-                    <HassuStack direction={["column", "column", "row"]}>
-                      <Chip
-                        icon={<HeadphonesIcon />}
-                        clickable
-                        label="Live-tilaisuus verkossa"
-                        variant="outlined"
-                        onDelete={() => console.log("lul")}
-                        deleteIcon={<HassuBadge badgeContent={7} color={"primary"} />}
-                      />
-                      <Chip
-                        icon={<LocationCityIcon />}
-                        clickable
-                        label="Fyysinen tilaisuus"
-                        variant="outlined"
-                        onDelete={() => console.log("lul")}
-                        deleteIcon={<HassuBadge badgeContent={2} color={"primary"} />}
-                      />
-                      <Chip
-                        icon={<LocalPhoneIcon />}
-                        clickable
-                        label="Soittoaika"
-                        variant="outlined"
-                        onDelete={() => console.log("lul")}
-                        deleteIcon={<HassuBadge badgeContent={1} color={"primary"} />}
-                      />
-                    </HassuStack>
-                    <Section>
-                      <TextInput
-                        label="Tilaisuuden nimi *"
-                        // {...register2("syy", { required: "Palautuksen syy täytyy antaa" })}
-                        // error={errors2.syy}
-                        maxLength={200}
-                      ></TextInput>
-                      <HassuStack direction={["column", "column", "row"]}>
-                        <DatePicker label="Päivämäärä *"></DatePicker>
-                        <TextInput label="Alkaa *" maxLength={6}></TextInput>
-                        <TextInput label="Päättyy *" maxLength={6}></TextInput>
-                      </HassuStack>
-                      <HassuGrid cols={{ lg: 3 }}>
-                        <Select
-                          addEmptyOption
-                          options={[
-                            { label: "Teams", value: "TEAMS" },
-                            { label: "Zoom", value: "ZOOM" },
-                          ]}
-                          label="Käytettävä palvelu *"
-                        />
-                      </HassuGrid>
-                      <TextInput label="Linkki tilaisuuteen *" maxLength={200}></TextInput>
-                      <p>Linkki tilaisuuteen julkaistaan palvelun julkisella puolella kaksi (2) tuntia ennen tilaisuuden alkamista.</p>
-                      <Button>Poista</Button>
-                    </Section>
-                  </HassuStack>
-                  <HassuStack
-                    direction={["column", "column", "row"]}
-                    justifyContent={[undefined, undefined, "flex-end"]}
-                    paddingTop={"1rem"}
-                  >
-                    <Button primary onClick={() => console.log("tallenna")}>
-                      Tallenna
-                    </Button>
-                    <Button
-                      onClick={(e) => {
-                        setOpenVuorovaikutustilaisuus(false);
-                        e.preventDefault();
-                      }}
-                    >
-                      Peruuta
-                    </Button>
-                  </HassuStack>
-                </form>
-              </div>
-            </SectionContent>
-          </Section>
-        </HassuDialog>
-      </div>
+      <VuorovaikutusDialog open={openVuorovaikutustilaisuus} windowHandler={setOpenVuorovaikutustilaisuus}></VuorovaikutusDialog>
       <HassuSpinner open={isFormSubmitting} />
     </>
   );
